Revert "feat: 유저 기록값 변경 API 추가"
This reverts commit c4ba21537786b2d94f21663067f78a817d09fd0f.

+++ user/src/main/kotlin/dailyquest/user/record/entity/UserRecord.kt
@@ -73,11 +73,11 @@ class UserRecord @JvmOverloads constructor(
     var goldUseAmount: Long = goldUseAmount
         protected set
 
-    fun recordGoldEarn(goldEarnAmount: Long) {
+    fun recordGoldEarn(goldEarnAmount: Int) {
         this.goldEarnAmount += goldEarnAmount
     }
 
-    fun recordGoldUse(goldUseAmount: Long) {
+    fun recordGoldUse(goldUseAmount: Int) {
         this.goldUseAmount += goldUseAmount
     }
